==> SPTT_MOB/01_GMM_Filter.R <==
rm(list = ls()); gc()
setwd("/your/workdir/")
library(data.table)
library(dplyr)
library(ggplot2)
library(tidyverse)
library(mclust)
library(Matrix)
library(RColorBrewer)

## load dge

dge = as.data.frame(fread(file = "./MOB_top5W_dge.txt.gz"))
dge = column_to_rownames(dge, "GENE")

# barcode_mapping.txt.gz is generated by Barcode_Degen workflow
coords=as.data.frame(fread(file = "./barcode_mapping.txt.gz")) 
colnames(coords) = c('raw_barcode', "degen_barcodes", "new_x", "new_y")
table(colnames(dge) %in% coords$degen_barcodes)

meta = unique(coords[which(coords$degen_barcodes %in% colnames(dge)), c("degen_barcodes", "new_x", "new_y")])


## UMI Model
umis = as.data.frame(Matrix::colSums(dge))
colnames(umis) = c("UMIs"); umis$degen_barcodes = rownames(umis)
meta = inner_join(meta, umis, by = "degen_barcodes")

GMM_umi_model = Mclust(meta$UMIs, G = 3, verbose = T)
meta$umi_gmm = as.character(GMM_umi_model$classification)
umi_gmm_avgr = aggregate(meta$UMIs, list(meta$umi_gmm), mean) 

## dist model
sp.loc = meta[, c("new_x", "new_y")]
sp.loc$new_x = as.integer(sp.loc[["new_x"]]); sp.loc$new_y = as.integer(sp.loc[["new_y"]])
nnDist = FNN::get.knnx(sp.loc[, c("new_x", "new_y")], sp.loc[, c("new_x", "new_y")], k = 20)

loc.mat = cbind(nnDist$nn.index, round(nnDist$nn.dist, 2))
mean(loc.mat[1, 22:dim(loc.mat)[2]])

average_dist = function(x, loc.mat)
{
  return(mean(loc.mat[x, 22:dim(loc.mat)[2]]))
}
nrows = 1:nrow(loc.mat)
beads_average_dist = lapply(nrows, average_dist, loc.mat)
beads_average_dist = do.call(c, beads_average_dist)
meta$average_dist = beads_average_dist
GMM_dist_model = Mclust(meta$average_dist, G = 3, verbose = T)
meta$dist_gmm = as.character(GMM_dist_model$classification)

meta.filter = meta %>% filter(dist_gmm != "3") %>% filter(UMIs >= 300) %>% filter(umi_gmm != "1")

# clean the margin
meta.filter = meta.filter %>% filter(new_y <= 7000 & new_x >= 3100)

# clean the margin 2nd Because there are still some scattered points around....
## dist model
sp.loc = meta.filter[, c("new_x", "new_y")]
sp.loc$new_x = as.integer(sp.loc[["new_x"]]); sp.loc$new_y = as.integer(sp.loc[["new_y"]])
nnDist = FNN::get.knnx(sp.loc[, c("new_x", "new_y")], sp.loc[, c("new_x", "new_y")], k = 20)

loc.mat = cbind(nnDist$nn.index, round(nnDist$nn.dist, 2))
mean(loc.mat[1, 22:dim(loc.mat)[2]])

average_dist = function(x, loc.mat)
{
  return(mean(loc.mat[x, 22:dim(loc.mat)[2]]))
}
nrows = 1:nrow(loc.mat)
beads_average_dist = lapply(nrows, average_dist, loc.mat)
beads_average_dist = do.call(c, beads_average_dist)
summary(beads_average_dist)
meta.filter$average_dist2 = beads_average_dist
GMM_dist_model = Mclust(meta.filter$average_dist2, G = 3, verbose = T)
meta.filter$dist_gmm2 = as.character(GMM_dist_model$classification)

meta.filter = meta.filter %>% filter(dist_gmm2 != "3")

######## filter dge
saveRDS(meta.filter, file = "./Barcode_level_meta_filter.rds")
dge_filter = dge[, meta.filter$degen_barcodes]
saveRDS(dge_filter, file = "./Barcode_level_dge_filter.rds")
